+++ Libraries/LibCore/SocketWindows.cpp
@@ -11,6 +11,7 @@
 #include <LibCore/System.h>
 
 #include <AK/Windows.h>
+#include <ws2tcpip.h>
 
 #define MSG_DONTWAIT 0x40
